scripts/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
python-pptx>=0.6.22
python-docx>=0.8.11
typing-extensions>=4.9.0